+++ qucore/js/src/main/scala/com/publicscript/qucore/EntityProjectileGrenade.scala
@@ -22,7 +22,7 @@ class EntityProjectileGrenade(apos:Vec3) extends Entity(apos) {
     this.f = if (this.on_ground) 5 else 0.5
   }
 
-  override def did_collide(axis: Double):Unit = {
+  override def did_collide(axis: Int):Unit = {
     if (axis != 1 || this.veloc.y < -128) {
       this.yaw += Math.random()
       this.play_sound(sfx_grenade_bounce)
