+++ qucore/js/src/main/scala/com/publicscript/qucore/EntityProjectileNail.scala
@@ -16,7 +16,7 @@ class EntityProjectileNail(apos:Vec3) extends Entity(apos) {
     this.draw_model()
   }
 
-  override def did_collide(axis: Double):Unit = {
+  override def did_collide(axis: Int):Unit = {
     this.kill()
     this.play_sound(sfx_nailgun_hit)
     this.spawn_particles(2, 80, model_explosion, 8, 0.4)
